修改excel导出，添加合计行，修改模式三计算方法

+++ back2/src/main/java/com/yupi/springbootinit/model/dto/tempmonitor/DailyEnergyConsumption.java
@@ -3,12 +3,17 @@ package com.yupi.springbootinit.model.dto.tempmonitor;
 import lombok.Data;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.Date;
 
 /**
  * 每日电能消耗数据 DTO
  * 对应 SQL 查询返回字段：deviceId、name、workshop、day、
  * energyConsumption、startEnergy、endEnergy、startTime、endTime
+ * 
+ * @author 每天十点睡
+ * @date 2026-02-09
+ * 修改：使用 BigDecimal 替代 Double，确保精确计算
  */
 @Data
 public class DailyEnergyConsumption implements Serializable {
@@ -28,13 +33,13 @@ public class DailyEnergyConsumption implements Serializable {
     private Date day;
 
     /** 当日用电量（kWh，按度数差值计算） */
-    private Double energyConsumption;
+    private BigDecimal energyConsumption;
 
     /** 日期开始时间以后最近记录的度数 */
-    private Double startEnergy;
+    private BigDecimal startEnergy;
 
     /** 日期结束时间（或当天最新）以后最近记录的度数 */
-    private Double endEnergy;
+    private BigDecimal endEnergy;
 
     /** 日期开始时间以后最近记录的时间 */
     private Date startTime;
