+++ back2/src/main/java/com/yupi/springbootinit/model/dto/tempmonitor/HourlyEnergyConsumption.java
@@ -3,12 +3,15 @@ package com.yupi.springbootinit.model.dto.tempmonitor;
 import lombok.Data;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.Date;
 
 /**
  * 每小时用电量数据DTO
  *
- * @author yupi
+ * @author 每天十点睡
+ * @date 2026-02-09
+ * 修改：使用 BigDecimal 替代 Double，确保精确计算
  */
 @Data
 public class HourlyEnergyConsumption implements Serializable {
@@ -38,17 +41,17 @@ public class HourlyEnergyConsumption implements Serializable {
     /**
      * 该小时的用电量（度数差值）
      */
-    private Double energyConsumption;
+    private BigDecimal energyConsumption;
 
     /**
      * 小时开始时间以后最近记录的度数
      */
-    private Double startEnergy;
+    private BigDecimal startEnergy;
 
     /**
      * 小时结束时间以后最近记录的度数
      */
-    private Double endEnergy;
+    private BigDecimal endEnergy;
 
     /**
      * 小时开始时间以后最近记录的时间
